Report game state
Serial is used by default but other modes could also
be added easily. For instance an exteral display can
be added via SPI or I2C. You can easily disable the
LED or BUTTON messages by selecting the appropriate
define.

LED messages should start with '*' followed by the
set of RGBY characters that are lit (others are
assumed to be off).

BUTTON messages should start with '#' and similarly
followed by the RGBY characters that are pressed.